api/requirements.txt
fastapi
uvicorn
google-generativeai
pypdf
python-multipart
python-dotenv
